+++ shareaza/DlgLanguage.cpp
@@ -1,7 +1,7 @@
 //
 // DlgLanguage.cpp
 //
-// Copyright (c) Shareaza Development Team, 2002-2012.
+// Copyright (c) Shareaza Development Team, 2002-2014.
 // This file is part of SHAREAZA (shareaza.sourceforge.net)
 //
 // Shareaza is free software; you can redistribute it
@@ -117,7 +117,7 @@ BOOL CLanguageDlg::OnInitDialog()
 	pScroll.cbSize	= sizeof(pScroll);
 	pScroll.fMask	= SIF_RANGE|SIF_PAGE|SIF_DISABLENOSCROLL;
 	pScroll.nMin	= 0;
-	pScroll.nMax	= ( m_pPaths.GetSize() / 3 ) + ( ( m_pPaths.GetSize() % 3 ) ? 1 : 0 );
+	pScroll.nMax	= ( (int)m_pPaths.GetSize() / 3 ) + ( ( (int)m_pPaths.GetSize() % 3 ) ? 1 : 0 );
 	pScroll.nPage	= m_nRows + 1;
 	SetScrollInfo( SB_VERT, &pScroll, TRUE );
